+++ ttcards/src/CardsMenuState.h
@@ -32,7 +32,7 @@ class CardsMenu: public nom::IState
     void Pause ( void );
     void Resume ( void );
 
-    void Update ( uint32_t elapsed_time );
+    void Update ( float elapsed_time );
     void Draw ( void *video_buffer );
 
   private:
